# app/swagger.yml
swagger: "2.0"
info:
  description: This is a documentation and playground for the siibra client endpoints
  version: "1.0.0"
  title: Siibra REST API
consumes:
  - "application/json"
produces:
  - "application/json"

basePath: "/api"

# Paths supported by the server application
paths:
  /receptors/fingerprint:
    get:
      operationId: "siibra_api.receptordata_fingerprint"
      tags:
        - "Features"
      summary: "Receptor fingerprint for region"
      description: "Get receptor fingerprint"
      parameters:
        - name: region
          in: query
          type: string
          description: Region name
          required: true
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get receptor fingerprint"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /receptors/profiles:
    get:
      operationId: "siibra_api.receptordata_profiles"
      tags:
        - "Features"
      summary: "Receptor profiles for region"
      description: "Get receptor profiles"
      parameters:
        - name: region
          in: query
          type: string
          description: Region name
          required: true
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get receptor profiles"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /receptors/autoradiographs:
    get:
      operationId: "siibra_api.receptordata_autoradiographs"
      tags:
        - "Features"
      summary: "Receptor autoradiographs for region"
      description: "Get receptor autoradiographs"
      parameters:
        - name: region
          in: query
          type: string
          description: Region name
          required: true
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get receptor autoradiographs"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /parcellations:
    get:
      operationId: "siibra_api.parcellations"
      tags:
        - "Parcellations"
      summary: "All parcellations for an atlas"
      description: "Return all available parcellations for selected atlas"
      parameters:
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get parcellations"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /spaces:
    get:
      operationId: "siibra_api.spaces"
      tags:
        - "Spaces"
      summary: "All spaces for an atlas"
      description: "Return all available spaces for selected atlas"
      parameters:
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get spaces"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /regions:
    get:
      operationId: "siibra_api.regions"
      tags:
        - "Regions"
      summary: "All regions for a parcellation"
      description: "Return all available regions for selected parcellation"
      parameters:
        - name: parcellation
          in: query
          type: string
          description: parcellation
          required: false
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      responses:
        200:
          description: "Successful get regions"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"
  /maps:
    get:
      operationId: "siibra_api.maps"
      tags:
        - "Spaces"
      summary: "Map for selected space"
      description: "Return a map for selected space"
      parameters:
        - name: space
          in: query
          type: string
          description: space
          required: true
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      produces: 
        - "application/octet-stream"          
      responses:
        200:
          description: "Successful get map"
          schema:
            type: "file"
  /templates:
    get:
      operationId: "siibra_api.templates"
      tags:
        - "Spaces"
      summary: "Template for selected space"
      description: "Return a template for selected space"
      parameters:
        - name: space
          in: query
          type: string
          description: space
          required: true
        - name: token
          in: query
          type: string
          description: Authentication token
          required: true
      produces: 
        - "application/octet-stream"
      responses:
        200:
          description: "Successful get template"
          schema:
            type: "file"
  /genes:
    get:
      operationId: "siibra_api.genes"
      tags:
        - "Features"
      summary: "Genes"
      description: "Return gene features"
      parameters:
        - name: region
          in: query
          type: string
          description: region
          required: true
        - name: gene
          in: query
          type: string
          description: gene
          required: true
      responses:
        200:
          description: "Successful get gene features"
          schema:
            type: "object"
            items:
              properties:
                content:
                  type: "object"